dlna: add verification of addresses
Verify the http service listening address and the ssdp server announcement
address to prevent accidental listening of ipv6 addresses that do not support
dlna and may be globally accessible.
Unlistened addresses on the interface will also be filtered out before the
ssdp announcement to avoid misleading other services in the broadcast domain.